Module atmega168_hal::atmega168

ATmega168 Peripheral access API for ATMEGA168 microcontrollers (generated using svd2rust v0.17.0 (5b392b4f3451 2020-08-15))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Modules

ac
Analog Comparator

adc
Analog-to-Digital Converter

cpu
CPU Registers

eeprom
EEPROM

exint
External Interrupts

fuse
Fuses

lockbit
Lockbits

portb
I/O Port

portc
I/O Port

portd
I/O Port

spi
Serial Peripheral Interface

tc0
Timer/Counter, 8-bit

tc1
Timer/Counter, 16-bit

tc2
Timer/Counter, 8-bit Async

twi
Two Wire Serial Interface

usart0
USART

wdt
Watchdog Timer

Structs

AC
Analog Comparator

ADC
Analog-to-Digital Converter

CPU
CPU Registers

EEPROM
EEPROM

EXINT
External Interrupts

FUSE
Fuses

LOCKBIT
Lockbits

PORTB
I/O Port

PORTC
I/O Port

PORTD
I/O Port

Peripherals
All the peripherals

SPI
Serial Peripheral Interface

TC0
Timer/Counter, 8-bit

TC1
Timer/Counter, 16-bit

TC2
Timer/Counter, 8-bit Async

TWI
Two Wire Serial Interface

USART0
USART

WDT
Watchdog Timer

Enums

Interrupt
Enumeration of all the interrupts