Module atmega168_hal::spi

Implementation of the Rust Embedded-HAL SPI FullDuplex trait for AVR.

The interface can be instantiated with the new method, and used directly or passed into a driver. Example usage:

// create SPI interface
let (mut spi, mut cs) = Spi::new(
    dp.SPI,// SPI peripheral
    pins.d13.into_output(&mut pins.ddr),// SCLK
    pins.d11.into_output(&mut pins.ddr),// MOSI output pin
    pins.d12.into_pull_up_input(&mut pins.ddr),// MISO input pin
    pins.d10.into_output(&mut pins.ddr),// CS pin
    Settings::default(),
);

// Send a byte
let sent = 0b10101010;
spi.send(sent).unwrap();
let response = spi.read().unwrap();

In the example above, all of the settings are left at the default. You can also instantiate a Settings object with the other options available.

Structs

ChipSelectPin
Wrapper for the CS pin

Settings
Settings to pass to Spi.

Spi
Behavior for a SPI interface.

Enums

DataOrder
Order of data transmission, either MSB first or LSB first

SerialClockRate
Oscillator Clock Frequency division options.